move finality_tester to a header file so it can be used by other tests; change clusetr to tester which makes more sense
